Fix class parameters
Values for the username, oauth_token, project_username, and
project_password parameters must be provided by the user. The empty
string is not a sane default and will cause broken deployments if used.
This change makes these parameters required. Infra is providing these
parameters in system-config so this change will not break Infra. If
downstream users are not providing these parameters their deployments
will not be functional, so this change should be safe for users already
using this class correctly.

The use of the projects parameter was removed in 15daf9a so it no
longer needs to be part of the parameter list. Infra is no longer using
the projects parameter in system-config so this change will not break
Infra. If downstream users still have this parameter as part of their
manifest, updating the module to this commit will break their puppet
runs, but not make any system changes.

Change-Id: Ic7d42f922ae436fe1b7b9bef21ca659ddedae432